+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/DashboardBackButton/DashboardBackButton.tsx
@@ -11,10 +11,12 @@ import {
 } from "metabase/query_builder/components/view/ViewHeader/ViewTitleHeader.styled";
 import { getDashboard } from "metabase/query_builder/selectors";
 
-interface DashboardBackButtonProps {
-  noLink?: boolean;
-  onClick?: () => void;
-}
+type DashboardBackButtonProps =
+  | { noLink?: false; onClick?: () => void }
+  | {
+      noLink: true;
+      onClick: () => void;
+    };
 
 export function DashboardBackButton({
   noLink,
